add tenant , delete tenant added,

+++ app/app/src/main/java/com/PGmitra/app/Repository/TenantRepo.java
@@ -3,5 +3,5 @@ package com.PGmitra.app.Repository;
 import com.PGmitra.app.Entity.Tenant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TenantRepo extends JpaRepository<Tenant, Integer> {
+public interface TenantRepo extends JpaRepository<Tenant, Long> {
 }
